re-organise controller tests
